For codon models, added check for very short branch lengths on the 3 "important" branches before entering into radius opt.  When these branches are very short but their MLE's are large, numerical instabilty can cause problems, including wildly incorrect lnL, d1, and d2.  Most importantly, the sign of d1 can be flipped, which causes obvious problems for NR optimization.  This is very rare, but hard to diagnose when it is happening.
Blens on these branches shorter than 1e-4 are now set to 1e-4.  This
aleviates the problem in all cases that I've seen.  It doesn't really
increase the number of opt passes much because the existing blens
tend to be crappy for those branches anyway, and if the MLE really is
very small (i.e. 1e-8) the NR function figures that out quickly already.